+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 - The canonical format (`-Tcanon`) produces output that is stable across
   multiple invocations. #2193
 
+- SVG output root dimensions no longer include point units. #867
+
 ## [16.1.0] – 2026-09-03
 
 ### Added